feat: text control web component (#142)
Reusable text control web component that can be used for input from controls. The text control doesn't construct an input element, but expects the consumer to slot in the input element:

```html
<oryx-text-control>
  <input></input>
</oryx-text-control>
```

With this setup, all native behaviour, attributes and events that come with the web platform will be available. 

The text control comes with some additional features:
- prefix icons, fill and custom content
- suffix icons, fill and custom content
- error messages

closes [FAAS-1329](https://spryker.atlassian.net/browse/FAAS-1329)