feat: wrap form controls in label element (#393)
We've not been wrapping the form control into a label element, which is causing potential accessibility concerns. 

There wasn't a good reason to do this, only it requires a little more code to ensure that the popover from the typeahead understand when there is label content, so that it can move the absolute position.

We've also introduced a slot for custom label content. 

closes [oryx-2065](https://spryker.atlassian.net/browse/FAAS-2065)

BREAKING CHANGES
- we've dropped the `<sup>*</sup>` for mandatory labels as we learned there was never an intention to do so