fix: only toggle the dropdown by spacebar in readonly mode (#398)
Toggling the dropdown that is used in typeahead and select components should only be offered with the space key when the input is not writable. That's the case for an readonly input or select element. 

While this fixes the original bug, another bug might arise at the horizon; whenever the spacebar is used inside client side filtering of typeahead options, there might be more expectations regarding filtering multipel words (separated by space).

closes [FAAS-2062](https://spryker.atlassian.net/browse/FAAS-2062)