fix: dynamic max height of popover depends on available space at bottom or bottom (#496)
When there's not enough space at the bottom, the popover is shown at the top. However, when the space is even smaller at the top, it will remain at the bottom. 
In both cases, the max-height of the popover is limited to the available space. 

fixes [FAAS-2187](https://spryker.atlassian.net/browse/FAAS-2187)