chore: make form controls mandatory (#535)
We've cleaned up the code base related to form controls (input, password, search, typeahead and select) so that we no longer need to check if a form control is available. 

closes [FAAS-2187](https://spryker.atlassian.net/browse/FAAS-2187)

BREAKING CHANGE: Form controls, such as `input`, `select` or `textarea`, are now mandatory. Whenever the control is not available, and error will be thrown. This makes our APIs cleaner, as we no longer need to use conditionals whenever we query for the form control. There's still a fallback `input` form control available in the main slot.