fix: placeholder option and interactive demo stories (#583)
We've fixed a number of issues with the interactive demo in storybook for the select component. 
We also introduced a couple of interactive stories that can be played in storybook. 

There are a couple of ambiguous things left in the demo story:
- a select component build with `<select>` options (useSelect=true) cannot be filtered. the combination of the `useSelect` and `filterStrategy` gives the impression that you can. 
- a select component build with `<input>`  and `oryx-option` elements (useSelect=false)  will always allow an empty value. The combination of `allowEmptyValue` and useSelect=false might be confusing. 

Additionally, a placeholder holder option is added for select components that use the native `<select>` structure.  Previously, this was not possible, but now we offer a mechanism to create an empty, hidden, first option that contains a placeholder text. 

fixes [FAAS-2576](https://spryker.atlassian.net/browse/FAAS-2576), [FAAS-2599](https://spryker.atlassian.net/browse/FAAS-2599), [FAAS-2812](https://spryker.atlassian.net/browse/FAAS-2812)

BREAKING_CHANGES:
- A new placeholder style is introduced for all form controls, using a CSS variable `--oryx-color-placeholder`.
- The padding for `<input />` and `<select>` based selects are now aligned. 
- When a placeholder is used, the select can be made `required`, so that the right placeholder style can be applied

The following declarative structure is required to use a placeholder for the `<select>`:

```html
<oryx-select>
  <select required>
    <option hidden value="">Placeholder text...</option>
  </select>
</oryx-select>
```