feat:  Rating Component (#643)
The rating component can be used to display an average rating (readonly) or set a rate. The rating component is used in a lot of places in a storefront, typically on the PLP and PDP, but also on carousels, in the facet filter navigation or on the overall site.

The rating component can be used with starts (default behaviour) or customised with SVGs or characters (i.e. emojis). 
The scale defaults to 5, but can be customised for an alternative scale such as 7. 

closes [FAAS-2673](https://spryker.atlassian.net/browse/FAAS-2673)