feat: avoid filtering when first character is a space (#736)
When the first character is a space, the user likely intends to open the select rather than filter by space. Hence, we prevent the default behaviour of the filter and keep the value empty. However, when a space is used when a filter is already applied (say "New "), we keep filtering, with the space included (so you can get "New York", "New Jersey", etc.).

closes [FAAS-3041](https://spryker.atlassian.net/browse/FAAS-3041)