set value when the option is selected (#738)
When an input is used to filter the select options, the input can change after the option has been selected. This feature was not working before, and resulted in a bug that blocks the user to reselect an option. This has been fixed, now the option can always be reselected. The events, however, will still not be dispatched again. 

closes [FAAS-3042](https://spryker.atlassian.net/browse/FAAS-3042)