fix: allow for overflow in card component (#756)
The modal component reuses the card component and requires to add scrollbars to the body of the card whenever the content is too large. This behaviour was baked into the card component, but this blocks the popover with select options to render outside the card component. We've therefor moved the overflow to the modal component. 

fixes [FAAS-3195](https://spryker.atlassian.net/browse/FAAS-3195)