feat: product and content controller (#765)
To avoid redundant code in each component, we introduce controllers to standardize the interaction with domain services. 


The product component controller depends on the `ProductComponentProperties` interface, that uses the product SKU. 

```ts
@property() sku?: string;
protected controller = new ProductComponentController(this);

...
() => {
  this.controller.product$
}
```


The component controller depends on the `ComponentProperties` interface, that uses the uid.  

```ts
@property() sku?: string;
protected controller = new ComponentController(this);

...
() => {
  this.controller.content$
}
```

closes [FAAS-3201](https://spryker.atlassian.net/browse/FAAS-3201)