fix: improve popover toggle controller focus loss logic (#935)
Basic popover toggle controller’s logic that check losing of focus is based on checking of document.activeElement, that works wrong when component with that controller is nested inside another custom component. Due to nuances of shadowDom work, document.activeElement can’t get the correct target from nested component. Focus passes to the parent component.

Second iteration of this fix. I've tried to change the approach and simplify the solution by simple check of ':focus-within' matching on host element. The bottle neck of this solution was that JSDOM doesn't support this selector. What was done:

- some sort of polyfill that simulates work of ':focus-within' selector in combination with elements 'matches' method call
- toggle controller focus loss logic was improved (the aim of this fix)
- dropdown component tests coverage was improved (due to adding :focus-within support that used inside it)

closes to FAAS-3572