fix: avoid initial filtered value on select (#980)
When `oryx-select` is configured with a so called filter, and there's an initial value, the value was used for filtering. This is an inconvenient  UX pattern and we change this so that only if the user types characters, we consider it a filter action. 

fixes [FAAS-3482](https://spryker.atlassian.net/browse/FAAS-3482)