feat: inline collapsible (#1289)
While building the cart totals, cart entries and checkout components, we discovered a new collapsible component: Inline collapsible. Inline collapsible has a different appearance and can use text button for the collapsible action. 
We've also introduced an alternative UI for the block collapsible for storefronts to support the checkout component. 

closes [HRZ-1037](https://spryker.atlassian.net/browse/HRZ-1037)