fix: apply margin to layout height (#1387)
When a margin is applied to the grid item, the (flex)width of this item should be decreased by the margin to ensure the layout is consistent. I.e. in a 4 column grid layout, where 1 of the items gets a 10px margin, we should still have 4 grid items of 25% width; the margin should be applied inside the column width. 

fixes [HRZ-1310](https://spryker.atlassian.net/browse/HRZ-1310)

Additionally, we added a dynamic SKU in the product card option, this must have stopped working.