fix: drop allowEmptyValue (#1410)
When oryx-select is setup to add an empty value programmatically (using allowEmptyValue), there's no way to tell if the empty option should be selected. The reason for this is that the original "select" option will select the first option by default in case there's no option selected. However, if this option is also selected based on the data, there's no way for oryx-select to know whether this done by data or by default behaviour. 

We therefor drop `allowEmptyValue` and force the host component to add an empty option manually. As part of this ticket we've refactor the usage in the form builder and ***. 

fixes [HRZ-1320](https://spryker.atlassian.net/browse/HRZ-1320)