fix: apply text overflow styles of chip to the its slot (#1431)
Current version of safari has a bug with re-rendering of content. Previously I made a conclusion that it was related to the paddings and custom elements. But in result when parent container has overflow: hidden this glitch is appearing. The solution is moving text overflow related styles to the nested element (in this case it is a slot)

Closes to HRZ-1386