fix: drop image load event (#1441)
The fetched property is currently blocking browsers to show an image while loading. This seems not needed, and is causing issues. Mainly on SSR as well as when progressive image are used. When an image fails to load, the broken image is removed by the error handler, the load handler is not playing a part here. 

We've dropped the onLoad event handler and associated logic. 

fixes [HRZ-1425](https://spryker.atlassian.net/browse/HRZ-1425)