fix: icon-button toggle siblings fix (#1457)
The simple fix that define the root node of component. Not the complete solution, because this won't work in case if icon-button will be a part of render of another web component that is slotted inside parent container. For such case will need to define strategy, maybe create a global register of our components that are presented on the page to simplify interaction with them

closes to HRZ-1317