fix: re-render of address list (#47)
Checked attribute sets only the default state and doesn't affect on property, but LIT is able to manages it under the hood and boolean checked attribute changes the prop. But for some cases, like deletion of currently checked radio button, this alignments doesn't work. To overcome this issue keyed directive is using for address list component, to trigger re-render and align attribute and property correctly

closes to HRZ-2237