chore: heading small screen appearance (#77)
We added appearance styles for small screens.

As part of this PR, we change the configuration of the screen sizes; screen sizes no longer will flow into larger screens, but are limited to the min AND max config of the breakpoint. As a result, we've been refactoring a few component styles which depends on the small screen size rules that was inherited by larger screen sizes originally.