feat(experience): configurable typography for compositions (#579)
The footer links must be of size `small`. In order to simplify this, we
introduced typography through composition styles.

closes: [HRZ-3187](https://spryker.atlassian.net/browse/HRZ-3187)

BREAKING CHANGE: we've moved the global font size of the html the root
of the document (using `ThemeMetaInitializer`, so that we can move our
typography font-size tokens to `rem` rather than `em`. Previously, all
typography was based on `em`.

[HRZ-3187]:
https://spryker.atlassian.net/browse/HRZ-3187?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ