feat(content): content image component (#585)
The content image component replaces the banner component. The banner
component was an component provided early on the project that was behind
a lot of concepts. With the configurable layout system there's no need
to provide a banner component, although we might bring a variation back
in the future. For now, complex banners can be created with a
combination of components (text, image, video, etc.) and layout.

closes: [HRZ-3216](https://spryker.atlassian.net/browse/HRZ-3216)

BREAKING CHANGE: The banner component is removed. 

[HRZ-3216]:
https://spryker.atlassian.net/browse/HRZ-3216?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ