refactor(ui): merge button and icon-button component (#618)
The button component has gone through some changes. The button now
supports an icon type, so that the icon-button component could be
dropped.

closes: [HRZ-3283](https://spryker.atlassian.net/browse/HRZ-3283)

BREAKING CHANGE: 
- The `oryx-icon-button` is removed, use `<oryx-button type="icon">`
going forward.
- Previously, it was mandatory to _slot in_ a `<button>` or `<a>`
element. While this is still supported, using the `slot="custom"`,
buttons will now create the button or link based on properties.
- The `type` property was previously used to indicate both the color and
style. This has been separated, so that you can define the `type` using
the `ButtonType` enum (`solid`, `outline`, `icon` or `text`) and the
color using the `ButtonColor` enum (`primary`, `neutral` or `error`)
separately.
- Previously the `outline` was set as a separate property, this property
is removed and moved into the `ButtonType` enum.
- The `size` property is now limited to the values of the `ButtonType`,
so that other sizes (`xs` and `xl`) cannot be used on the button
component.


[HRZ-3283]:
https://spryker.atlassian.net/browse/HRZ-3283?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

---------

Co-authored-by: pavel.gulvanskiy <pavlo.hulvanskiy@spryker.com>
Co-authored-by: Andrii Ievtukhov <andrii.ievtukhov@spryker.com>