fix(cart): ensure handling of string-based cart entry quantities (#804)
While connecting to a b2b backend in preparation of customer call, it
turns out that the cart entry quantity is of type string. So far, we’ve
been seeing the number type for quantity.

We’re not entirely sure if quantity is always going to be numeric, hence
we do not force the conversion of the response in the adapter layer. For
now, we’ll assume it’s a quantity in the component layer, but over time
we can change this when we have more clarity on the usage of string vs
number.

closes: [HRZ-90046](https://spryker.atlassian.net/browse/HRZ-90046)

[HRZ-90046]:
https://spryker.atlassian.net/browse/HRZ-90046?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ