feat(cart): separate cart heading component (#726)
The separation of the cart heading and the cart entries improves the
reusability of both elements and will help to add more features to the
cart heading going forward. Consequently, we introduced a checkout and
order heading.

The heading components are considered new features that will official
land in 2.0, but can already be used and applied automatically when the
`ORYX_FEATURE_VERSION` (min. `1.2`) is used.

closes: [HRZ-3262](https://spryker.atlassian.net/browse/HRZ-3244)

[HRZ-3262]:
https://spryker.atlassian.net/browse/HRZ-3262?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

---------

Co-authored-by: tobi-or-not-tobi <tobi-or-not-tobi@users.noreply.github.com>