feat(experience): carousel navigation (#856)
The carousel navigation can be used in combination with the carousel
layout. The layout can be used with a combination of items from the
light DOM or shadow DOM, either direct children or provided by nested
components (e.g. product list component).

The carousel navigation supports a number of features: 
- Dynamic calculation of slides.
- Navigation by item or by slide.
- (optional) arrow controls to navigate to the previous or next item or
slide. The controls disappear when they're not applicable at the
carousel state, e.g. when the first slide is active, the previous
control will disappear.
- (optional) slide indicator controls, allow you to navigate to
individual slides.
- Highlight indicators (gradually) while scrolling the carousel. Both
when manually scrolling as well as when navigating by the controls.
- (optional) allow for scrolling through the carousel by mouse.
- (optional) allow for scrolling through the carousel by touch.
- Instant vs smooth sliding behaviour.
- Indicator horizontal alignment: start, end, and center.
- Indicator vertical alignment: below, bottom.
- Vertical and horizontal carousels

closes: [HRZ-90099](https://spryker.atlassian.net/browse/HRZ-90099)


[HRZ-90099]:
https://spryker.atlassian.net/browse/HRZ-90099?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

---------

Co-authored-by: Oleksii Bilan <alexey.belan@spryker.com>
Co-authored-by: pavel.gulvanskiy <pavlo.hulvanskiy@spryker.com>