feat: configurable forced color mode (#907)
Oryx supports both dark and light mode colors, leveraging the device
color mode or by an application color mode selector. However, some
applications are not ready to go in dark mode. This is also the case for
the Fulfillment App.
To still leverage the rich color palette, we introduce an option on the
Oryx App (root) component to control forcing the light vs dark mode.

This feature can be enabled by providing a configuration on the
application:

```ts
appBuilder()
  ...
  .withOptions({
    'oryx-app': { colorMode: ColorMode.Light },
  })
  .create();
```

closes: [HRZ-90485](https://spryker.atlassian.net/browse/HRZ-90485)


[HRZ-90485]:
https://spryker.atlassian.net/browse/HRZ-90485?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ