feat(experience): header category navigation (#882)
The header navigation contains some of the categories available in the
assortment, which differ from b2c to b2b. The category is build with
standard components, such compositions and link components, and the
layout is provided by a new navigation layout. The navigation layout
supports both horizontal or vertical layouts.

The navigation layouts adds a number of (opinionated) CSS variables that
are used in the styles of the descendants content link components:
- `--oryx-link-icon-size`
- `--oryx-link-width`
- `--oryx-link-color`
- `--oryx-link-decoration`
- `--oryx-link-padding`
- `--oryx-link-hover-shadow`
- `--oryx-link-active-shadow`
- `--oryx-link-current-shadow`

closes: [HRZ-2519](https://spryker.atlassian.net/browse/HRZ-2519)

[HRZ-2519]:
https://spryker.atlassian.net/browse/HRZ-2519?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ