feat(cart): coupon component (#950)
Coupon can be applied to the cart to get discounts for some specific product. More than one coupon can be applied to a cart, each coupon have a validation date that's it presented to the user.

There is a validation on the coupon input to give feedback to the user. Coupons are both used for guest users and authenticated users.

Cart service and adapter updated to get  coupons and add a new one

Closes HRZ-2852