test: fix fa e2e tests after mock removal (#956)
* refactored 85% of tests and supporting files
* removed useless tests from suites
* split suites in E2E (integration folder) and components tests (they
should not be a part of the E2E check, but should be covered by
unit/component tests)

---
There are 2 sources of flakiness which I can't handle:

* when you log in - in 50% of cases you are stuck on the loading screen
because some backend requests never occur. I'm 99% sure this is a
front-end issue
* when you create a new order and go to back office ui -> in 50% of
cases 3 minutes (!!!) is not enough to process the order and change the
status so that it can be moved to ready for picking. It is purely a
BE-related topic, we can't fix it. From what I know, a job that changes
statuses runs each minute and I have no idea why 3 minutes on a test
server (which no one uses) is insufficient.

---
If checks are fine, and CR is approved -> I'll merge development in this
branch to check if it works after all refactorings we made (don't want
to mess the code with possible merge conflict resolutions)

closes: https://spryker.atlassian.net/browse/HRZ-89888

---------

Co-authored-by: Alexander Kovalenko <aleksander.kovalenko@spryker.com>