chore(experience): fixed additional stuff with preview outline (#1005)
Components that have a `display` set to `contents` failed to have a good outline, as the component does not expose a bounding box that can be used to calculate the outline. We've worked around this problem by iterating over all children to find the outer most top, right, bottom and left dimensions.