fix(ui): notifications's i18n content support (#1044)
i18n directive uses signals under the hood, that requires first initialisation of the value (usually it is undefined). This handled via life-circle hooks inside components, but for cases like notification center where content for notification is passed to the push method inside config, the signal is not re-init the value. To overcome such limitation additional support of I18nContent was added to the ui notification center

- NotificationContent interface now supports I18nContent ({token?: string, context?: InferI18nContext, raw?: string})
- I18nContent interface was added to the '@spryker-oryx/utilities'
- TextResource interface of @spryker-oryx/i18n package was deprecated. Use I18nContent instead
- Notification and NotificationContent interfaces were moved from @spryker-oryx/ui/notification to '@spryker-oryx/ui/notification-center', and deprecated for NotificationComponent

closes: HRZ-90684